win x86: a few trivial compile fixes when GYP_DEFINES=target_arch=ia32
(CL to add x86 bots to waterfall in progress too.)

R=mark@chromium.org
BUG=crashpad:49

Review URL: https://codereview.chromium.org/1325173002 .

+++ test/win/win_child_process.cc
@@ -54,7 +54,7 @@ bool GetSwitch(const char* switch_name, std::string* value) {
 
   std::string switch_name_with_equals(switch_name);
   switch_name_with_equals += "=";
-  for (size_t i = 1; i < num_args; ++i) {
+  for (int i = 1; i < num_args; ++i) {
     const wchar_t* arg = args[i];
     std::string arg_as_utf8 = base::UTF16ToUTF8(arg);
     if (arg_as_utf8.compare(
